web/backend/requirements.txt
# Vaulytica Web Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5.0.0
python-multipart==0.0.6

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Type Checking (dev)
mypy==1.7.1
types-python-jose==3.3.4.8
types-passlib==1.7.7.13
